feat: support Gateway API 1.6.0
Bump sigs.k8s.io/gateway-api from v1.3.0 to v1.6.0. This pulls in
k8s.io v0.36 and controller-runtime v0.24, which require Go 1.26, so
the go directive and CI Go version are bumped accordingly.

Adapt to the Gateway API v1.6 changes:
- Frontend mTLS validation moved from the per-listener TLS config
  (listener.tls.frontendValidation) to the Gateway level
  (spec.tls.frontend). Resolve the validation that applies to each
  HTTPS listener by port (PerPort overrides Default) in the gateway
  translator, indexer and status/reconcile paths.
- listener.TLS type renamed GatewayTLSConfig -> ListenerTLSConfig.
- HTTPCORSFilter.AllowCredentials is now *bool.
- Policy status/target types (PolicyStatus, PolicyAncestorStatus,
  LocalPolicyTargetReferenceWithSectionName, Policy* condition/reason
  constants) graduated to apis/v1; switch usages off v1alpha2.
- PortNumber is now an alias of int32; drop redundant conversions.

Adapt to controller-runtime v0.24:
- Webhook builder: NewWebhookManagedBy(mgr).For(obj) -> the generic
  NewWebhookManagedBy[runtime.Object](mgr, obj); WithValidator now
  takes the typed Validator[T] and CustomValidator/webhook.CustomValidator
  are deprecated aliases of Validator[runtime.Object].

Rewrite the conformance suite for the v1.6 conformance module API
(RunConformanceWithOptions, report written via ReportOutputPath).

Build golangci-lint from source with the local toolchain and pin the
lint target Go version, since the released binary cannot type-check a
Go 1.26 module.

+++ internal/adc/translator/grpcroute.go
@@ -205,7 +205,7 @@ func (t *Translator) TranslateGRPCRoute(tctx *provider.TranslateContext, grpcRou
 				kind = string(*backend.Kind)
 			}
 			if backend.Port != nil {
-				port = int32(*backend.Port)
+				port = *backend.Port
 			}
 			namespace := string(*backend.Namespace)
 			name := string(backend.Name)
@@ -312,7 +312,7 @@ func (t *Translator) TranslateGRPCRoute(tctx *provider.TranslateContext, grpcRou
 		// Collect unique listener ports for port-based routing.
 		listenerPorts := make(map[int32]struct{})
 		for _, listener := range tctx.Listeners {
-			listenerPorts[int32(listener.Port)] = struct{}{}
+			listenerPorts[listener.Port] = struct{}{}
 		}
 
 		if t.shouldInjectServerPortVars(tctx.RouteParentRefs, listenerPorts) {
